pyproject.toml
[project]
name = "py-claude-sandbox"
version = "0.1.0"
description = "Python implementation of Claude Code Sandbox - run Claude in isolated Podman containers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docker>=7.0.0",
    "typer>=0.9.0",
    "nicegui>=1.4.0",
    "GitPython>=3.1.0",
    "pydantic>=2.5.0",
    "watchdog>=3.0.0",
    "psutil>=5.9.0",
    "cryptography>=41.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-docker>=2.0.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
py-claude-sandbox = "py_claude_sandbox.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "PL"]
ignore = ["E501", "PLR2004"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as an async test",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]
